- work on book controller - configured hibernate search

+++ src/main/java/mk/ukim/finki/emt/model/jpa/FileEmbeddable.java
@@ -15,5 +15,5 @@ public class FileEmbeddable {
 
   public String contentType;
 
-  public Long size;
+  public int size;
 }
